refactor: option controller 분리에 따라 통합 테스트도 분리
